\name{lmmlasso-package}
\alias{lmmlasso-package}
\docType{package}
\title{lmmlasso}
\description{Fits (gaussian) linear mixed-effects models with lasso
  penalty for the fixed effects.}
\details{
\tabular{ll}{
Package: \tab lmmlasso\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2010-08-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

This is an early test version.
}

\author{
Juerg Schelldorfer

Maintainer: Juerg Schelldorfer <schell@stat.math.ethz.ch>
}
\references{
J. Schelldorfer, P. Buehlmann and S. van de Geer (2011), Estimation for High-Dimensional Linear Mixed-Effects
                  Models Using \eqn{\ell_1}-penalization, arXiv preprint 1002.3784v2
}
\keyword{package}
